+++ src/active/dkgsessionhandler.cpp
@@ -7,17 +7,12 @@
 #include <active/dkgsession.h>
 #include <active/masternode.h>
 #include <evo/deterministicmns.h>
-#include <llmq/blockprocessor.h>
 #include <llmq/debug.h>
 #include <llmq/dkgsession.h>
 #include <llmq/net_quorum.h>
-#include <llmq/options.h>
-#include <llmq/utils.h>
-#include <masternode/meta.h>
 
 #include <deploymentstatus.h>
 #include <logging.h>
-#include <util/thread.h>
 #include <validation.h>
 
 namespace llmq {
@@ -71,23 +66,9 @@ void ActiveDKGSessionHandler::UpdatedBlockTip(const CBlockIndex* pindexNew)
              params.name, quorumIndex, currentHeight, pQuorumBaseBlockIndex->nHeight, std23::to_underlying(oldPhase), std23::to_underlying(phase));
 }
 
-void ActiveDKGSessionHandler::StartThread(CConnman& connman, PeerManager& peerman)
+uint256 ActiveDKGSessionHandler::GetCurrentQuorumHash() const
 {
-    if (phaseHandlerThread.joinable()) {
-        throw std::runtime_error("Tried to start an already started ActiveDKGSessionHandler thread.");
-    }
-
-    m_thread_name = strprintf("llmq-%d-%d", std23::to_underlying(params.type), quorumIndex);
-    phaseHandlerThread = std::thread(&util::TraceThread, m_thread_name.c_str(),
-                                     [this, &connman, &peerman] { PhaseHandlerThread(connman, peerman); });
-}
-
-void ActiveDKGSessionHandler::StopThread()
-{
-    stopRequested = true;
-    if (phaseHandlerThread.joinable()) {
-        phaseHandlerThread.join();
-    }
+    return WITH_LOCK(cs_phase_qhash, return quorumHash);
 }
 
 std::pair<QuorumPhase, uint256> ActiveDKGSessionHandler::GetPhaseAndQuorumHash() const
@@ -115,9 +96,6 @@ bool ActiveDKGSessionHandler::InitNewQuorum(gsl::not_null<const CBlockIndex*> pQ
     LogPrintf("ActiveDKGSessionHandler::%s -- height[%d] quorum initialization OK for %s qi[%d]\n", __func__, pQuorumBaseBlockIndex->nHeight, params.name, quorumIndex);
     return true;
 }
-
-class AbortPhaseException : public std::exception {
-};
 
 void ActiveDKGSessionHandler::WaitForNextPhase(std::optional<QuorumPhase> curPhase, QuorumPhase nextPhase,
                                                const uint256& expectedQuorumHash, const WhileWaitFunc& shouldNotWait) const
@@ -249,336 +227,6 @@ void ActiveDKGSessionHandler::HandlePhase(QuorumPhase curPhase, QuorumPhase next
     WaitForNextPhase(curPhase, nextPhase, expectedQuorumHash, runWhileWaiting);
 
     LogPrint(BCLog::LLMQ_DKG, "ActiveDKGSessionHandler::%s -- %s qi[%d] - done, curPhase=%d, nextPhase=%d\n", __func__, params.name, quorumIndex, std23::to_underlying(curPhase), std23::to_underlying(nextPhase));
-}
-
-// returns a set of NodeIds which sent invalid messages
-template<typename Message>
-std::unordered_set<NodeId> BatchVerifyMessageSigs(CDKGSession& session, const std::vector<std::pair<NodeId, std::shared_ptr<Message>>>& messages)
-{
-    if (messages.empty()) {
-        return {};
-    }
-
-    std::unordered_set<NodeId> ret;
-    bool revertToSingleVerification = false;
-
-    CBLSSignature aggSig;
-    std::vector<CBLSPublicKey> pubKeys;
-    std::vector<uint256> messageHashes;
-    Uint256HashSet messageHashesSet;
-    pubKeys.reserve(messages.size());
-    messageHashes.reserve(messages.size());
-    bool first = true;
-    for (const auto& [nodeId, msg] : messages) {
-        auto member = session.GetMember(msg->proTxHash);
-        if (!member) {
-            // should not happen as it was verified before
-            ret.emplace(nodeId);
-            continue;
-        }
-
-        if (first) {
-            aggSig = msg->sig;
-        } else {
-            aggSig.AggregateInsecure(msg->sig);
-        }
-        first = false;
-
-        auto msgHash = msg->GetSignHash();
-        if (!messageHashesSet.emplace(msgHash).second) {
-            // can only happen in 2 cases:
-            // 1. Someone sent us the same message twice but with differing signature, meaning that at least one of them
-            //    must be invalid. In this case, we'd have to revert to single message verification nevertheless
-            // 2. Someone managed to find a way to create two different binary representations of a message that deserializes
-            //    to the same object representation. This would be some form of malleability. However, this shouldn't be
-            //    possible as only deterministic/unique BLS signatures and very simple data types are involved
-            revertToSingleVerification = true;
-            break;
-        }
-
-        pubKeys.emplace_back(member->dmn->pdmnState->pubKeyOperator.Get());
-        messageHashes.emplace_back(msgHash);
-    }
-    if (!revertToSingleVerification) {
-        if (aggSig.VerifyInsecureAggregated(pubKeys, messageHashes)) {
-            // all good
-            return ret;
-        }
-
-        // are all messages from the same node?
-        bool nodeIdsAllSame = std::adjacent_find( messages.begin(), messages.end(), [](const auto& first, const auto& second){
-            return first.first != second.first;
-        }) == messages.end();
-
-        // if yes, take a short path and return a set with only him
-        if (nodeIdsAllSame) {
-            ret.emplace(messages[0].first);
-            return ret;
-        }
-        // different nodes, let's figure out who are the bad ones
-    }
-
-    for (const auto& [nodeId, msg] : messages) {
-        if (ret.count(nodeId)) {
-            continue;
-        }
-
-        auto member = session.GetMember(msg->proTxHash);
-        bool valid = msg->sig.VerifyInsecure(member->dmn->pdmnState->pubKeyOperator.Get(), msg->GetSignHash());
-        if (!valid) {
-            ret.emplace(nodeId);
-        }
-    }
-    return ret;
-}
-
-static void RelayInvToParticipants(const CDKGSession& session, const CConnman& connman, PeerManager& peerman,
-                                   const CInv& inv)
-{
-    CDKGLogger logger(session, __func__, __LINE__);
-    std::stringstream ss;
-    const auto& relayMembers = session.RelayMembers();
-    for (const auto& r : relayMembers) {
-        ss << r.ToString().substr(0, 4) << " | ";
-    }
-    logger.Batch("RelayInvToParticipants inv[%s] relayMembers[%d] GetNodeCount[%d] GetNetworkActive[%d] "
-                 "HasMasternodeQuorumNodes[%d] for quorumHash[%s] forMember[%s] relayMembers[%s]",
-                 inv.ToString(), relayMembers.size(), connman.GetNodeCount(ConnectionDirection::Both),
-                 connman.GetNetworkActive(),
-                 connman.HasMasternodeQuorumNodes(session.GetType(), session.BlockIndex()->GetBlockHash()),
-                 session.BlockIndex()->GetBlockHash().ToString(), session.ProTx().ToString().substr(0, 4), ss.str());
-
-    std::stringstream ss2;
-    connman.ForEachNode([&](const CNode* pnode) {
-        if (pnode->qwatch ||
-            (!pnode->GetVerifiedProRegTxHash().IsNull() && (relayMembers.count(pnode->GetVerifiedProRegTxHash()) != 0))) {
-            peerman.PushInventory(pnode->GetId(), inv);
-        }
-
-        if (pnode->GetVerifiedProRegTxHash().IsNull()) {
-            logger.Batch("node[%d:%s] not mn", pnode->GetId(), pnode->m_addr_name);
-        } else if (relayMembers.count(pnode->GetVerifiedProRegTxHash()) == 0) {
-            ss2 << pnode->GetVerifiedProRegTxHash().ToString().substr(0, 4) << " | ";
-        }
-    });
-    logger.Batch("forMember[%s] NOTrelayMembers[%s]", session.ProTx().ToString().substr(0, 4), ss2.str());
-    logger.Flush();
-}
-
-template <typename Message>
-bool ProcessPendingMessageBatch(const CConnman& connman, CDKGSession& session, CDKGPendingMessages& pendingMessages,
-                                PeerManager& peerman, size_t maxCount)
-{
-    auto msgs = pendingMessages.PopAndDeserializeMessages<Message>(maxCount);
-    if (msgs.empty()) {
-        return false;
-    }
-
-    std::vector<std::pair<NodeId, std::shared_ptr<Message>>> preverifiedMessages;
-    preverifiedMessages.reserve(msgs.size());
-
-    for (const auto& p : msgs) {
-        const NodeId &nodeId = p.first;
-        if (!p.second) {
-            LogPrint(BCLog::LLMQ_DKG, "%s -- failed to deserialize message, peer=%d\n", __func__, nodeId);
-            peerman.Misbehaving(nodeId, 100);
-            continue;
-        }
-        bool ban = false;
-        if (!session.PreVerifyMessage(*p.second, ban)) {
-            if (ban) {
-                LogPrint(BCLog::LLMQ_DKG, "%s -- banning node due to failed preverification, peer=%d\n", __func__, nodeId);
-                peerman.Misbehaving(nodeId, 100);
-            }
-            LogPrint(BCLog::LLMQ_DKG, "%s -- skipping message due to failed preverification, peer=%d\n", __func__, nodeId);
-            continue;
-        }
-        preverifiedMessages.emplace_back(p);
-    }
-    if (preverifiedMessages.empty()) {
-        return true;
-    }
-
-    auto badNodes = BatchVerifyMessageSigs(session, preverifiedMessages);
-    if (!badNodes.empty()) {
-        for (auto nodeId : badNodes) {
-            LogPrint(BCLog::LLMQ_DKG, "%s -- failed to verify signature, peer=%d\n", __func__, nodeId);
-            peerman.Misbehaving(nodeId, 100);
-        }
-    }
-
-    for (const auto& p : preverifiedMessages) {
-        const NodeId &nodeId = p.first;
-        if (badNodes.count(nodeId)) {
-            continue;
-        }
-        const std::optional<CInv> inv = session.ReceiveMessage(*p.second);
-        if (inv) {
-            RelayInvToParticipants(session, connman, peerman, *inv);
-        }
-    }
-
-    return true;
-}
-
-static void AddQuorumProbeConnections(const Consensus::LLMQParams& llmqParams, CConnman& connman, CMasternodeMetaMan& mn_metaman,
-                               const CSporkManager& sporkman, const UtilParameters& util_params,
-                               const CDeterministicMNList& tip_mn_list, const uint256& myProTxHash)
-{
-    assert(mn_metaman.IsValid());
-
-    if (!IsQuorumPoseEnabled(llmqParams.type, sporkman)) {
-        return;
-    }
-
-    auto members = utils::GetAllQuorumMembers(llmqParams.type, util_params);
-    auto curTime = GetTime<std::chrono::seconds>().count();
-
-    Uint256HashSet probeConnections;
-    for (const auto& dmn : members) {
-        if (dmn->proTxHash == myProTxHash) {
-            continue;
-        }
-        auto lastOutbound = mn_metaman.GetLastOutboundSuccess(dmn->proTxHash);
-        if (curTime - lastOutbound < 10 * 60) {
-            // avoid re-probing nodes too often
-            continue;
-        }
-        probeConnections.emplace(dmn->proTxHash);
-    }
-
-    if (!probeConnections.empty()) {
-        if (LogAcceptDebug(BCLog::LLMQ)) {
-            std::string debugMsg = strprintf("%s -- adding masternodes probes for quorum %s:\n", __func__,
-                                             util_params.m_base_index->GetBlockHash().ToString());
-            for (const auto& c : probeConnections) {
-                auto dmn = tip_mn_list.GetValidMN(c);
-                if (!dmn) {
-                    debugMsg += strprintf("  %s (not in valid MN set anymore)\n", c.ToString());
-                } else {
-                    debugMsg += strprintf("  %s (%s)\n", c.ToString(),
-                                          dmn->pdmnState->netInfo->GetPrimary().ToStringAddrPort());
-                }
-            }
-            LogPrint(BCLog::NET_NETCONN, debugMsg.c_str()); /* Continued */
-        }
-        connman.AddPendingProbeConnections(probeConnections);
-    }
-}
-void ActiveDKGSessionHandler::HandleDKGRound(CConnman& connman, PeerManager& peerman)
-{
-    WaitForNextPhase(std::nullopt, QuorumPhase::Initialized);
-
-    pendingContributions.Clear();
-    pendingComplaints.Clear();
-    pendingJustifications.Clear();
-    pendingPrematureCommitments.Clear();
-    uint256 curQuorumHash = WITH_LOCK(cs_phase_qhash, return quorumHash);
-
-    const CBlockIndex* pQuorumBaseBlockIndex = WITH_LOCK(::cs_main,
-                                                         return m_chainman.m_blockman.LookupBlockIndex(curQuorumHash));
-
-    if (!pQuorumBaseBlockIndex || !InitNewQuorum(pQuorumBaseBlockIndex)) {
-        // should actually never happen
-        WaitForNewQuorum(curQuorumHash);
-        throw AbortPhaseException();
-    }
-
-    m_dkgdbgman.UpdateLocalSessionStatus(params.type, quorumIndex, [&](CDKGDebugSessionStatus& status) {
-        bool changed = status.phase != QuorumPhase::Initialized;
-        status.phase = QuorumPhase::Initialized;
-        return changed;
-    });
-
-    if (params.is_single_member()) {
-        auto finalCommitment = curSession->FinalizeSingleCommitment();
-        if (!finalCommitment.IsNull()) { // it can be null only if we are not member
-            if (auto inv_opt = m_qblockman.AddMineableCommitment(finalCommitment); inv_opt.has_value()) {
-                peerman.RelayInv(inv_opt.value());
-            }
-        }
-        WaitForNextPhase(QuorumPhase::Initialized, QuorumPhase::Contribute, curQuorumHash);
-        return;
-    }
-
-    const auto tip_mn_list = m_dmnman.GetListAtChainTip();
-    llmq::EnsureQuorumConnections(params, connman, m_sporkman, {m_dmnman, m_qsnapman, m_chainman, pQuorumBaseBlockIndex},
-                                   tip_mn_list, curSession->ProTx(), /*is_masternode=*/true, m_quorums_watch);
-    if (curSession->AreWeMember()) {
-        AddQuorumProbeConnections(params, connman, m_mn_metaman, m_sporkman,
-                                         {m_dmnman, m_qsnapman, m_chainman, pQuorumBaseBlockIndex}, tip_mn_list,
-                                         curSession->ProTx());
-    }
-
-    WaitForNextPhase(QuorumPhase::Initialized, QuorumPhase::Contribute, curQuorumHash);
-
-    // Contribute
-    auto fContributeStart = [this]() {
-        if (auto qc = curSession->Contribute(); qc) {
-            pendingContributions.PushOwnPendingMessage(*qc);
-        }
-    };
-    auto fContributeWait = [this, &connman, &peerman] {
-        return ProcessPendingMessageBatch<CDKGContribution>(connman, *curSession, pendingContributions, peerman, 8);
-    };
-    HandlePhase(QuorumPhase::Contribute, QuorumPhase::Complain, curQuorumHash, 0.05, fContributeStart, fContributeWait);
-
-    // Complain
-    auto fComplainStart = [this, &connman]() {
-        if (auto qc = curSession->VerifyAndComplain(connman); qc) {
-            pendingComplaints.PushOwnPendingMessage(*qc);
-        }
-    };
-    auto fComplainWait = [this, &connman, &peerman] {
-        return ProcessPendingMessageBatch<CDKGComplaint>(connman, *curSession, pendingComplaints, peerman, 8);
-    };
-    HandlePhase(QuorumPhase::Complain, QuorumPhase::Justify, curQuorumHash, 0.05, fComplainStart, fComplainWait);
-
-    // Justify
-    auto fJustifyStart = [this]() {
-        if (auto qj = curSession->VerifyAndJustify(); qj) {
-            pendingJustifications.PushOwnPendingMessage(*qj);
-        }
-    };
-    auto fJustifyWait = [this, &connman, &peerman] {
-        return ProcessPendingMessageBatch<CDKGJustification>(connman, *curSession, pendingJustifications, peerman, 8);
-    };
-    HandlePhase(QuorumPhase::Justify, QuorumPhase::Commit, curQuorumHash, 0.05, fJustifyStart, fJustifyWait);
-
-    // Commit
-    auto fCommitStart = [this]() {
-        if (auto qc = curSession->VerifyAndCommit(); qc) {
-            pendingPrematureCommitments.PushOwnPendingMessage(*qc);
-        }
-    };
-    auto fCommitWait = [this, &connman, &peerman] {
-        return ProcessPendingMessageBatch<CDKGPrematureCommitment>(connman, *curSession, pendingPrematureCommitments,
-                                                                   peerman, 8);
-    };
-    HandlePhase(QuorumPhase::Commit, QuorumPhase::Finalize, curQuorumHash, 0.1, fCommitStart, fCommitWait);
-
-    auto finalCommitments = curSession->FinalizeCommitments();
-    for (const auto& fqc : finalCommitments) {
-        if (auto inv_opt = m_qblockman.AddMineableCommitment(fqc); inv_opt.has_value()) {
-            peerman.RelayInv(inv_opt.value());
-        }
-    }
-}
-
-void ActiveDKGSessionHandler::PhaseHandlerThread(CConnman& connman, PeerManager& peerman)
-{
-    while (!stopRequested) {
-        try {
-            LogPrint(BCLog::LLMQ_DKG, "ActiveDKGSessionHandler::%s -- %s qi[%d] - starting HandleDKGRound\n", __func__, params.name, quorumIndex);
-            HandleDKGRound(connman, peerman);
-        } catch (AbortPhaseException& e) {
-            m_dkgdbgman.UpdateLocalSessionStatus(params.type, quorumIndex, [&](CDKGDebugSessionStatus& status) {
-                status.statusBits.aborted = true;
-                return true;
-            });
-            LogPrint(BCLog::LLMQ_DKG, "ActiveDKGSessionHandler::%s -- %s qi[%d] - aborted current DKG session\n", __func__, params.name, quorumIndex);
-        }
-    }
 }
 
 bool ActiveDKGSessionHandler::GetContribution(const uint256& hash, CDKGContribution& ret) const
